Fix logging implicit sign conversions
The warnings were being hidden by the use of -isystem to include
frameworks/native/include.

Bug: 31752268
Test: m -j
(cherry picked from commit 152c3b749a8ec634afc66dddef7d33130472aeb1)

+++ libs/gui/BufferQueueProducer.cpp
@@ -905,7 +905,8 @@ status_t BufferQueueProducer::queueBuffer(int slot,
                 static_cast<uint32_t>(mCore->mQueue.size()),
                 mCore->mFrameCounter + 1);
 
-        ATRACE_INT(mCore->mConsumerName.string(), mCore->mQueue.size());
+        ATRACE_INT(mCore->mConsumerName.string(),
+                static_cast<int32_t>(mCore->mQueue.size()));
         mCore->mOccupancyTracker.registerOccupancyChange(mCore->mQueue.size());
 
         // Take a ticket for the callback functions
